+++ ITStudy/Models/Vouchers.cs
@@ -7,7 +7,7 @@ public class Vouchers
     public Student? Student { get; set; }
     public Courses? Course { get; set; }
     public string Title { get; set; } = string.Empty;
-    public int Discount { get; set; } = 0;
+    public long Discount { get; set; } = 0;
     public long? CourseId { get; set; } = null;
     public DateTime? Public_Date { get; set; } = DateTime.UtcNow.ToUniversalTime();
     public DateTime? Expire_Date { get; set; } = DateTime.UtcNow.ToUniversalTime();
